Refactor(auth): 로그인 응답, 요청 DTO 폴더구조 변경
